# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoCRM API",
    "description": "Autotherm project lifecycle system. Money is integer minor units with an explicit currency; quantities are decimal strings. PATCH: omit a field to keep it, send null to clear it.",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/emails": {
      "get": {
        "tags": [
          "email"
        ],
        "operationId": "email_list",
        "parameters": [
          {
            "name": "order_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "lead_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "partner_id",
            "in": "query",
            "description": "Includes mail about the partner's orders and leads.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EmailStatus"
            }
          },
          {
            "name": "attention",
            "in": "query",
            "description": "Only failed and needs-review mail.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_EmailSummary"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "email"
        ],
        "operationId": "email_send",
        "requestBody": {
          "description": "Subject/body override the template. Unresolved variables are rejected.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComposeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailMessage"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/emails/preview": {
      "post": {
        "tags": [
          "email"
        ],
        "operationId": "email_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComposeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preview"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/emails/{id}": {
      "get": {
        "tags": [
          "email"
        ],
        "operationId": "email_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailMessage"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/emails/{id}/cancel": {
      "post": {
        "tags": [
          "email"
        ],
        "operationId": "email_cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cancelled"
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/emails/{id}/retry": {
      "post": {
        "tags": [
          "email"
        ],
        "operationId": "email_retry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Re-queued. Only failed or needs-review mail."
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/email-templates": {
      "get": {
        "tags": [
          "email"
        ],
        "operationId": "email_list_templates",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_EmailTemplate"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "email"
        ],
        "operationId": "email_create_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailTemplate"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/email-templates/variables": {
      "get": {
        "tags": [
          "email"
        ],
        "operationId": "email_variables",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_TemplateVariable"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/email-templates/{id}": {
      "patch": {
        "tags": [
          "email"
        ],
        "operationId": "email_update_template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTemplate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailTemplate"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/email-suppressions": {
      "get": {
        "tags": [
          "email"
        ],
        "operationId": "email_list_suppressions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_Suppression"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "email"
        ],
        "operationId": "email_add_suppression",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSuppression"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Added"
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/email-suppressions/{email}": {
      "delete": {
        "tags": [
          "email"
        ],
        "operationId": "email_remove_suppression",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComposeRequest": {
        "type": "object",
        "required": [
          "to"
        ],
        "properties": {
          "order_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "lead_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "partner_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "to": {
            "type": "string"
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "template_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject": {
            "type": [
              "string",
              "null"
            ]
          },
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "attachment_document_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "CreateTemplate": {
        "type": "object",
        "required": [
          "key",
          "name",
          "subject",
          "body"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
...
      "EmailStatus": {
        "type": "string",
        "enum": [
          "queued",
          "sending",
          "sent",
          "failed",
          "cancelled",
          "needs_review"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "Items_EmailSummary": {
        "type": "object",
        "description": "The list envelope: `{\"items\": [...]}`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "List view: everything except the bodies.",
              "required": [
                "id",
                "trigger",
                "is_automatic",
                "to_address",
                "subject",
                "status",
                "queued_at",
                "send_after"
              ],
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "order_id": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "lead_id": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "partner_id": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "blocker_id": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "template_key": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "trigger": {
                  "type": "string"
                },
                "is_automatic": {
                  "type": "boolean"
                },
                "sent_by": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "sent_by_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "to_address": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "status": {
                  "$ref": "#/components/schemas/EmailStatus"
                },
                "error": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "queued_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "send_after": {
                  "type": "string",
                  "format": "date-time"
                },
                "sent_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "PatchTemplate": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject": {
            "type": [
              "string",
              "null"
            ]
          },
          "body": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Preview": {
        "type": "object",
        "required": [
          "to",
          "cc",
          "subject",
          "body_text",
          "body_html",
          "unresolved",
          "recipient_suppressed",
          "attachments"
        ],
        "properties": {
          "to": {
            "type": "string"
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "string"
          },
          "body_text": {
            "type": "string"
          },
          "body_html": {
            "type": "string"
          },
          "unresolved": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recipient_suppressed": {
            "type": "boolean",
            "description": "Manual mail may still go to a suppressed address; the UI should warn."
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentRef"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "autocrm_session"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "bearer": []
    }
  ],
...
}